+++ core-web/.eslintrc.json
@@ -26,10 +26,19 @@
             "files": ["*.ts", "*.tsx"],
             "extends": ["plugin:@nrwl/nx/typescript"],
             "rules": {
-                "@typescript-eslint/keyword-spacing": ["error", { "after": true }],
+                "@typescript-eslint/keyword-spacing": [
+                    "error",
+                    {
+                        "after": true
+                    }
+                ],
                 "@typescript-eslint/padding-line-between-statements": [
                     "error",
-                    { "blankLine": "always", "prev": "*", "next": "return" },
+                    {
+                        "blankLine": "always",
+                        "prev": "*",
+                        "next": "return"
+                    },
                     {
                         "blankLine": "always",
                         "prev": ["case", "default", "block-like", "export", "for", "function"],
@@ -43,8 +52,20 @@
                         "argsIgnorePattern": "^_"
                     }
                 ],
-                "one-var": ["error", { "var": "never", "let": "never", "const": "never" }],
-                "no-console": ["error", { "allow": ["warn", "error"] }]
+                "one-var": [
+                    "error",
+                    {
+                        "var": "never",
+                        "let": "never",
+                        "const": "never"
+                    }
+                ],
+                "no-console": [
+                    "error",
+                    {
+                        "allow": ["warn", "error"]
+                    }
+                ]
             }
         },
         {
