+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/components/dot-experiments-empty-experiments/dot-experiments-empty-experiments.component.html
@@ -4,7 +4,7 @@
     <p-button
         *ngIf="showButton"
         [label]="'experimentspage.add.new.experiment' | dm"
-        (onClick)="addNewBtn.emit()"
+        (onClick)="addNew.emit()"
         data-testId="add-btn"
     ></p-button>
 </div>
