+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/components/dot-experiments-empty-experiments/dot-experiments-empty-experiments.component.ts
@@ -16,5 +16,5 @@ export class DotExperimentsEmptyExperimentsComponent {
     showButton = true;
 
     @Output()
-    addNewBtn = new EventEmitter<void>();
+    addNew = new EventEmitter<void>();
 }
