+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/components/dot-experiments-list-skeleton/dot-experiments-list-skeleton.component.html
@@ -1,26 +1,57 @@
-<div class="row">
-    <p-skeleton width="14rem" height="2rem"></p-skeleton>
-    <p-skeleton width="10rem" height="2rem"></p-skeleton>
-</div>
-<div class="row" style="margin-top: 1rem">
-    <p-skeleton width="8rem" height="1.5rem"></p-skeleton>
-</div>
+<div class="grid formgrid" style="margin-top: 1rem">
+    <div class="field col-12">
+        <div class="flex justify-content-between mt-3">
+            <!-- Filter  -->
+            <p-skeleton height="2rem" width="15rem"></p-skeleton>
+            <!-- New experiment  -->
+            <p-skeleton height="2rem" width="13rem"></p-skeleton>
+        </div>
+    </div>
 
-<div class="row line">
-    <p-skeleton width="30rem" height="1rem"></p-skeleton>
-    <p-skeleton width="8rem" height="1rem"></p-skeleton>
-    <p-skeleton width="8rem" height="1rem"></p-skeleton>
-    <p-skeleton width="8rem" height="1rem"></p-skeleton>
-</div>
-<div class="row line">
-    <p-skeleton width="30rem" height="1rem"></p-skeleton>
-    <p-skeleton width="8rem" height="1rem"></p-skeleton>
-    <p-skeleton width="8rem" height="1rem"></p-skeleton>
-    <p-skeleton width="8rem" height="1rem"></p-skeleton>
-</div>
-<div class="row line">
-    <p-skeleton width="30rem" height="1rem"></p-skeleton>
-    <p-skeleton width="8rem" height="1rem"></p-skeleton>
-    <p-skeleton width="8rem" height="1rem"></p-skeleton>
-    <p-skeleton width="8rem" height="1rem"></p-skeleton>
+    <!-- Status  -->
+    <div class="field col-12 mt-4">
+        <p-skeleton height="1.5rem" width="8rem"></p-skeleton>
+    </div>
+
+    <!-- Row  -->
+    <div class="field col-7">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+    <div class="field col-2">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+    <div class="field col-2">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+    <div class="field col-1">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+
+    <!-- Row  -->
+    <div class="field col-7">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+    <div class="field col-2">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+    <div class="field col-2">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+    <div class="field col-1">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+
+    <!-- Row  -->
+    <div class="field col-7">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+    <div class="field col-2">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+    <div class="field col-2">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
+    <div class="field col-1">
+        <p-skeleton height="1rem" width="100%"></p-skeleton>
+    </div>
 </div>
