--- core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/components/dot-experiments-list-skeleton/dot-experiments-list-skeleton.component.scss
@@ -1,20 +1,0 @@
-@use "variables" as *;
-
-:host {
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    width: 100%;
-    overflow: auto;
-    .row {
-        display: flex;
-        flex-direction: row;
-        width: 100%;
-        justify-content: space-between;
-        margin-bottom: 15px;
-        &.line {
-            justify-content: start;
-            column-gap: 3rem;
-        }
-    }
-}
