--- core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/components/dot-experiments-list-skeleton/dot-experiments-list-skeleton.component.spec.ts
@@ -1,22 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DotExperimentsListSkeletonComponent } from './dot-experiments-list-skeleton.component';
-
-describe('DotExperimentsSkeletonComponent', () => {
-    let component: DotExperimentsListSkeletonComponent;
-    let fixture: ComponentFixture<DotExperimentsListSkeletonComponent>;
-
-    beforeEach(async () => {
-        await TestBed.configureTestingModule({
-            declarations: [DotExperimentsListSkeletonComponent]
-        }).compileComponents();
-
-        fixture = TestBed.createComponent(DotExperimentsListSkeletonComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
