+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/components/dot-experiments-list-skeleton/dot-experiments-list-skeleton.component.ts
@@ -3,7 +3,13 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 @Component({
     selector: 'dot-experiments-list-skeleton',
     templateUrl: './dot-experiments-list-skeleton.component.html',
-    styleUrls: ['./dot-experiments-list-skeleton.component.scss'],
+    styles: [
+        `
+            :host {
+                width: 100%;
+            }
+        `
+    ],
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DotExperimentsListSkeletonComponent {}
