+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/components/dot-experiments-list-table/dot-experiments-list-table.component.scss
@@ -2,98 +2,97 @@
 
 :host {
     .dot-experiment-list-table {
-        &__wrapper {
-            display: flex;
-            flex-direction: column;
-            gap: 20px;
-            margin-bottom: 2rem;
+        display: flex;
+        flex-direction: column;
+        gap: $spacing-4;
+        margin-bottom: $spacing-5;
+    }
+
+    //ng-table overwritte
+    ::ng-deep .p-datatable {
+        border-collapse: separate;
+        border-spacing: 0;
+
+        tr:last-child td:first-child {
+            border-bottom-left-radius: $border-radius-md;
         }
-    }
-}
 
-//ng-table overwritte
-::ng-deep .p-datatable {
-    border-collapse: separate;
-    border-spacing: 0;
+        td:last-child {
+            border-bottom-right-radius: $border-radius-md;
+        }
 
-    tr:last-child td:first-child {
-        border-bottom-left-radius: 4px;
-    }
-
-    td:last-child {
-        border-bottom-right-radius: 4px;
-    }
-
-    th,
-    td {
-        font-size: $font-size-medium;
-        font-weight: normal;
-    }
-
-    .p-datatable-thead > tr > th,
-    .p-datatable-tbody > tr > td {
-        background: transparent;
-    }
-
-    .p-datatable-thead {
-        tr > th {
-            height: 40px;
-            padding: 0;
-            color: $th-text-color-experiments;
+        th,
+        td {
             font-size: $font-size-medium;
             font-weight: normal;
-            border: 1px solid $table-border-color-experiments;
-            border-width: 0 0 1px 0;
-            font-weight: 500;
-
-            &:first-child {
-                padding-left: 20px;
-            }
-
-            &:last-child {
-                padding-right: 0;
-            }
         }
-    }
 
-    .p-datatable-header {
-        border: none;
-        color: $gray;
-        background: transparent;
-    }
+        .p-datatable-thead > tr > th,
+        .p-datatable-tbody > tr > td {
+            background: transparent;
+        }
 
-    .p-datatable-wrapper {
-        border-radius: 4px;
-        border: 1px solid $table-border-color-experiments;
-
-        .p-datatable-tbody {
-            tr {
+        .p-datatable-thead {
+            tr > th {
                 height: 40px;
-                border: 1px solid $table-border-color-experiments;
+                padding: 0;
+                color: $gray;
+                font-size: $font-size-medium;
+                font-weight: normal;
+                border: 1px solid $gray-lighter;
                 border-width: 0 0 1px 0;
-                transition: background-color $basic-speed ease;
+                font-weight: 500;
+
+                &:first-child {
+                    padding-left: 20px;
+                }
 
                 &:last-child {
-                    border-width: 0 0 0 0;
+                    padding-right: 0;
                 }
+            }
+        }
 
-                &:hover {
-                    background-color: #e0e9f6;
-                }
+        .p-datatable-header {
+            border: none;
+            color: $gray;
+            background: transparent;
+        }
 
-                td {
+        .p-datatable-wrapper {
+            border-radius: $border-radius-md;
+            border: 1px solid $gray-lighter;
+
+            .p-datatable-tbody {
+                tr {
                     height: 40px;
-                    padding: 0;
-                    border: none;
-
-                    &:first-child {
-                        padding-left: 20px;
-                    }
+                    border: 1px solid $gray-lighter;
+                    border-width: 0 0 1px 0;
+                    transition: background-color $basic-speed ease;
+                    background: transparent;
 
                     &:last-child {
-                        justify-content: flex-end;
-                        display: flex;
-                        padding-right: 1rem;
+                        border-width: 0 0 0 0;
+                    }
+
+                    &:hover {
+                        background-color: $tr-hover-color;
+                    }
+
+                    td {
+                        height: 40px;
+                        padding: 0;
+                        border: none;
+
+                        &:first-child {
+                            padding-left: $spacing-3;
+                        }
+
+                        &:last-child {
+                            justify-content: flex-end;
+                            display: flex;
+                            padding-right: $spacing-3;
+                        }
                     }
                 }
             }
