--- core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/components/dot-experiments-status-filter/dot-experiments-status-filter.component.scss
@@ -1,4 +1,0 @@
-:host ::ng-deep .p-multiselect {
-    min-width: 10rem;
-    width: 13rem;
-}
