+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/components/dot-experiments-status-filter/dot-experiments-status-filter.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
-import { DotExperimentStatusList } from '@portlets/dot-experiments/shared/models/dot-experiments.model';
+import { DotExperimentStatusList } from '@portlets/dot-experiments/shared/models/dot-experiments-constants';
 
 interface optionItem {
     label: string;
@@ -10,7 +10,6 @@ interface optionItem {
 @Component({
     selector: 'dot-experiments-status-filter',
     templateUrl: './dot-experiments-status-filter.component.html',
-    styleUrls: ['./dot-experiments-status-filter.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DotExperimentsStatusFilterComponent {
@@ -20,5 +19,5 @@ export class DotExperimentsStatusFilterComponent {
     options: optionItem[];
 
     @Output()
-    checkedOptions = new EventEmitter<DotExperimentStatusList[]>();
+    switch = new EventEmitter<DotExperimentStatusList[]>();
 }
