+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-list/dot-experiments-list.component.ts
@@ -1,52 +1,64 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { DotExperimentsStore } from '@portlets/dot-experiments/shared/stores/dot-experiments-store.service';
 import { Observable } from 'rxjs';
-import { DotExperiment, ExperimentsStatusList } from '../shared/models/dot-experiments.model';
+import { DotExperiment } from '../shared/models/dot-experiments.model';
+import { ExperimentsStatusList } from '@portlets/dot-experiments/shared/models/dot-experiments-constants';
 import { DotMessagePipe } from '@dotcms/app/view/pipes';
-
-interface VmListExperiments {
-    isLoading: boolean;
-    experiments: DotExperiment[];
-    experimentsFiltered: { [key: string]: DotExperiment[] };
-    experimentsFilterStatus: Array<string>;
-}
+import { ActivatedRoute } from '@angular/router';
+import {
+    DotExperimentsListStore,
+    VmListExperiments
+} from '@portlets/dot-experiments/dot-experiments-list/store/dot-experiments-list-store.service';
 
 @Component({
     selector: 'dot-experiments-list',
     templateUrl: './dot-experiments-list.component.html',
     styleUrls: ['./dot-experiments-list.component.scss'],
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    providers: [DotMessagePipe]
+    providers: [DotMessagePipe],
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DotExperimentsListComponent implements OnInit {
-    vmListExperiments$: Observable<VmListExperiments> = this.experimentsStore.vmExperimentList$;
+    vm$: Observable<VmListExperiments> = this.dotExperimentsListStore.vm$;
 
-    statusOptions = ExperimentsStatusList.map(({ label, value }) => {
+    statusOptionList = ExperimentsStatusList.map(({ label, value }) => {
         return { value, label: this.dotMessagePipe.transform(label) };
     });
 
     constructor(
-        private readonly experimentsStore: DotExperimentsStore,
-        private readonly dotMessagePipe: DotMessagePipe
+        private readonly dotExperimentsListStore: DotExperimentsListStore,
+        private readonly dotMessagePipe: DotMessagePipe,
+        private readonly route: ActivatedRoute
     ) {}
 
     ngOnInit() {
-        this.experimentsStore.loadExperiments();
+        const pageId = this.route.parent?.parent?.parent.snapshot.params.pageId;
+        this.dotExperimentsListStore.loadExperiments(pageId);
     }
 
     selectedStatusFilter($event: Array<string>) {
-        this.experimentsStore.setExperimentsFilterStatus($event);
+        this.dotExperimentsListStore.setFilterStatus($event);
     }
 
+    /**
+     * Add experiment
+     * @param experiment
+     */
     addNewExperiment() {
         // To implement
     }
 
-    archiveExperiment(experimentId: string) {
-        this.experimentsStore.archiveExperiment(experimentId);
+    /**
+     * Archive selected experiment
+     * @param experiment
+     */
+    archiveExperiment(experiment: DotExperiment) {
+        this.dotExperimentsListStore.archiveExperiment(experiment);
     }
 
-    deleteExperiment(experimentId: string) {
-        this.experimentsStore.deleteExperiment(experimentId);
+    /**
+     * Delete selected experiment
+     * @param experiment
+     */
+    deleteExperiment(experiment: DotExperiment) {
+        this.dotExperimentsListStore.deleteExperiment(experiment);
     }
 }
