+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-shell/dot-experiments-shell.component.html
@@ -1,9 +1,6 @@
-<ng-container *ngIf="vmShell$ | async as vm">
-    <dot-experiments-header
-        [title]="vm.currentTitlePage"
-        (goBack)="goBack()"
-    ></dot-experiments-header>
-    <dot-loading-indicator [show]="vm.status" fullscreen="true"></dot-loading-indicator>
+<ng-container *ngIf="vm$ | async as vm">
+    <dot-experiments-header [title]="vm.pageTitle" (goBack)="goBack()"></dot-experiments-header>
 </ng-container>
 
 <router-outlet></router-outlet>
+<p-toast></p-toast>
