+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-shell/dot-experiments-shell.component.ts
@@ -1,43 +1,47 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { DotExperimentsStore } from '../shared/stores/dot-experiments-store.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { DotExperimentsService } from '../shared/services/dot-experiments.service';
 import { Observable } from 'rxjs';
-import { LoadingState } from '@portlets/shared/models/shared-models';
-import { RenderedPageExperiments } from '@portlets/dot-experiments/shared/models/dot-experiments.model';
 import { DotPageRenderState } from '@portlets/dot-edit-page/shared/models';
+import { ComponentStore } from '@ngrx/component-store';
+import { RenderedPageExperiments } from '@portlets/dot-experiments/shared/models/dot-experiments.model';
 
-interface VmShellExperiments {
-    status: LoadingState;
-    currentTitlePage: string;
+export interface DotExperimentsState {
+    pageId: string;
+    pageTitle: string;
 }
 
 @Component({
     selector: 'dot-experiments-shell',
     templateUrl: 'dot-experiments-shell.component.html',
     styleUrls: ['./dot-experiments-shell.component.scss'],
-    providers: [DotExperimentsStore, DotExperimentsService],
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DotExperimentsShellComponent implements OnInit {
-    vmShell$: Observable<VmShellExperiments> = this.experimentsStore.vmShell$;
+    readonly vm$: Observable<DotExperimentsState> = this.componentStore.state$;
 
     constructor(
-        private readonly experimentsStore: DotExperimentsStore,
+        private readonly componentStore: ComponentStore<DotExperimentsState>,
         private readonly route: ActivatedRoute,
         private readonly router: Router
     ) {}
 
     ngOnInit() {
-        this.experimentsStore.initStore(this._getResolverExperimentsData(this.route));
+        const page = this._getResolverExperimentsData();
+        this.componentStore.setState({
+            pageId: page.identifier,
+            pageTitle: page.title
+        });
     }
 
+    /**
+     * Back to Edit Page / Content
+     */
     goBack() {
         this.router.navigate(['edit-page/content'], { queryParamsHandling: 'preserve' });
     }
 
-    private _getResolverExperimentsData(route: ActivatedRoute): RenderedPageExperiments {
-        const { page } = route.parent?.parent.snapshot.data?.content as DotPageRenderState;
+    private _getResolverExperimentsData(): RenderedPageExperiments {
+        const { page } = this.route.parent?.parent.snapshot.data?.content as DotPageRenderState;
 
         return {
             identifier: page.identifier,
