+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/shared/models/dot-experiments.model.ts
@@ -1,3 +1,7 @@
+import {
+    DotExperimentStatusList,
+    TrafficProportionTypes
+} from '@portlets/dot-experiments/shared/models/dot-experiments-constants';
 import { DotPage } from '@models/dot-page/dot-page.model';
 
 export interface DotExperiment {
@@ -19,37 +23,6 @@ export type RenderedPageExperiments = Pick<DotPage, 'title' | 'identifier'>;
 
 export type GroupedExperimentByStatus = Partial<Record<DotExperimentStatusList, DotExperiment[]>>;
 
-export enum DotExperimentStatusList {
-    RUNNING = 'RUNNING',
-    SCHEDULED = 'SCHEDULED',
-    ENDED = 'ENDED',
-    DRAFT = 'DRAFT',
-    ARCHIVED = 'ARCHIVED'
-}
-
-export const ExperimentsStatusList = [
-    {
-        label: 'experimentspage.experiment.status.draft',
-        value: DotExperimentStatusList.DRAFT
-    },
-    {
-        label: 'experimentspage.experiment.status.running',
-        value: DotExperimentStatusList.RUNNING
-    },
-    {
-        label: 'experimentspage.experiment.status.ended',
-        value: DotExperimentStatusList.ENDED
-    },
-    {
-        label: 'experimentspage.experiment.status.archived',
-        value: DotExperimentStatusList.ARCHIVED
-    },
-    {
-        label: 'experimentspage.experiment.status.scheduled',
-        value: DotExperimentStatusList.SCHEDULED
-    }
-];
-
 interface TrafficProportion {
     percentages: {
         [index: string]: number;
@@ -60,9 +33,4 @@ interface TrafficProportion {
 interface RangeOfDateAndTime {
     startDate: Date;
     endDate: Date;
-}
-
-export const enum TrafficProportionTypes {
-    SPLIT_EVENLY = 'SPLIT_EVENLY',
-    CUSTOM_PERCENTAGES = 'CUSTOM_PERCENTAGES'
 }
