+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/shared/ui/experiments-header/dot-experiments-ui-header.component.html
@@ -1,2 +1,4 @@
-<a (click)="goBack.emit(true)"><dot-icon name="arrow_back_ios" size="24"></dot-icon></a>
-<span>{{ title }}</span>
+<a (click)="goBack.emit(true)" data-testId="goback-button">
+    <dot-icon name="arrow_back_ios" size="24"></dot-icon>
+</a>
+<span data-testId="title">{{ title }}</span>
