+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/shared/ui/experiments-header/dot-experiments-ui-header.component.spec.ts
@@ -1,20 +1,41 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { DotExperimentsUiHeaderComponent } from './dot-experiments-ui-header.component';
+import { byTestId, createComponentFactory, Spectator } from '@ngneat/spectator';
+import { DotIconComponent } from '@dotcms/ui';
 
-xdescribe('ExperimentsHeaderComponent', () => {
-    let component: DotExperimentsUiHeaderComponent;
-    let fixture: ComponentFixture<DotExperimentsUiHeaderComponent>;
+describe('ExperimentsHeaderComponent', () => {
+    let spectator: Spectator<DotExperimentsUiHeaderComponent>;
+    let dotIconComponent: DotIconComponent | null;
 
-    beforeEach(async () => {
-        await TestBed.configureTestingModule({}).compileComponents();
-
-        fixture = TestBed.createComponent(DotExperimentsUiHeaderComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
+    const createComponent = createComponentFactory({
+        component: DotExperimentsUiHeaderComponent
     });
 
-    it('should create', () => {
-        expect(component).toBeTruthy();
+    beforeEach(() => {
+        spectator = createComponent();
+    });
+
+    it('should has a title rendered', () => {
+        const title = 'My title';
+        spectator.setInput('title', title);
+        expect(spectator.query(byTestId('title'))).toHaveText(title);
+    });
+
+    it('should has a dotIcon rendered', () => {
+        dotIconComponent = spectator.query(DotIconComponent);
+
+        expect(dotIconComponent).toExist();
+    });
+
+    it('should trigger the go Back Output', () => {
+        let output;
+        const backLinkIcon: HTMLAnchorElement = spectator.query(byTestId('goback-button'));
+
+        spectator.output('goBack').subscribe((result) => {
+            output = result;
+        });
+
+        backLinkIcon.click();
+
+        expect(output).toEqual(true);
     });
 });
