+++ core-web/libs/dot-primeng-theme-styles/src/scss/_variables.scss
@@ -19,6 +19,7 @@ $gray-bg: #f1f3f4;
 $gray-hover: #eee;
 $deleted-content-color: #fdb8c0;
 $new-content-color: #ddffdd;
+$tr-hover-color: #e0e9f6;
 
 $white_rgb: var(--color-white_rgb);
 $brand-primary: var(--color-main);
@@ -54,8 +55,6 @@ $button-shadow-focus: 0 0 0 4px rgba(var(--color-sec_rgb), 0.2);
 // Tables
 $th-text-color: $gray-light;
 $table-border-color: $gray-light;
-$table-border-color-experiments: $gray-lighter;
-$th-text-color-experiments: $gray;
 
 // Spacing
 $spacing-0: 0;
@@ -77,6 +76,7 @@ $success: $green;
 
 // Basic stuffs
 $border-radius: 2px;
+$border-radius-md: 4px;
 $field-height: 36px;
 $field-tiny-height: 28px;
 $field-disabled-opacity: 0.7;
@@ -155,5 +155,7 @@ $form-width: 400px;
 
 // CONTENT TYPES
 $content-type-field-height: 74px;
-
 $dot-editor-size: 16px;
+
+// CONFIRM POPUP
+$confirm-popup-column-gap: 5px;
